combined the Login and Register page into a component and update the App.js with the component name to display on the page
