+++ lowes-capstone/package.json
@@ -7,6 +7,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "autoprefixer": "^10.4.14",
     "axios": "^1.3.6",
     "mdb-react-ui-kit": "^6.0.0",
     "react": "^18.2.0",
