+++ lowes-capstone/src/Pages/Login.jsx
@@ -1,33 +1,5 @@
-// import React, { useState } from "react";
-// import Footer from "../Components/footer";
-
-
-// export const Login = (props) => {
-//     const [email, setEmail] = useState('');
-//     const [pass, setPass] = useState('');
-
-//     const handleSubmit = (e) => {
-//         e.preventDefault();
-//         console.log(email);
-//     }
-
-//     return (
-//         <div className="auth-form-container">
-//             <h2>Login</h2>
-//             <form className="login-form" onSubmit={handleSubmit}>
-//                 <label htmlFor="email">Email</label>
-//                 <input value={email} onChange={(e) => setEmail(e.target.value)}type="email" placeholder="youremail@gmail.com" id="email" name="email" />
-//                 <label htmlFor="password">Password</label>
-//                 <input value={pass} onChange={(e) => setPass(e.target.value)} type="password" placeholder="********" id="password" name="password" />
-//                 <button type="submit">Log In</button>
-//             </form>
-//             <button className="link-btn" onClick={() => props.onFormSwitch('register')}>Don't have an account? Register here.</button>
-//             <Footer />
-//         </div>
-//     )
-// }
 import React, { useState } from "react";
-import Footer from "../Components/footer";
+import "../Login.css";
 
 export const Login = (props) => {
   const [email, setEmail] = useState("");
@@ -39,7 +11,7 @@ export const Login = (props) => {
   };
 
   return (
-    <>
+      <>
       <div className="auth-form-container">
         <h2>Login</h2>
         <form className="login-form" onSubmit={handleSubmit}>
@@ -70,7 +42,6 @@ export const Login = (props) => {
           Don't have an account? Register here.
         </button>
       </div>
-        <Footer />
     </>
   );
 };
