+++ lowes-capstone/src/Pages/Register.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import zxcvbn from "zxcvbn"; //library used to test the strength of the password the client creates
-import Footer from "../Components/footer";
 
 // used OOP(Object Oriented Programming)
 //create a class that extends the React.Component class.
@@ -102,7 +101,6 @@ class Register extends React.Component {
         >
           Already have an account? Login here.
         </button>
-        <Footer />
       </div>
     );
   }
